Changed to paging source Didn't Implement Changing to one item per page for scrolling to position UI Revamp Bug : UI blank when imageview back pressed in SDK 22,23

+++ app/src/main/java/com/akhilasdeveloper/marsroverphotos/db/table/photo/key/DisplayPhotosTable.kt
@@ -7,5 +7,5 @@ import androidx.room.Entity
 data class DisplayPhotosTable(
     val roverName: String,
     val date: Long,
-    val photoID : Int
+    val photoID : Long
 )
